pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.4",
    "langchain-community>=0.4.1",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.16",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.6",
    "pypdf>=6.6.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
